\name{scmaster}
\alias{scmaster}

\title{
Simple and useful functions for supply chain management
}
\description{
  An R package for supply chain optimization and forecasting.  
  It includes functions for ABC analysis, Economic Order Quantity (EOQ),  
  safety stock calculation, and transportation optimization using linear programming.  
  The package provides essential tools for inventory management, logistics,  
  and demand planning, ensuring efficient decision-making in supply chain operations.
  }

\author{
Dari Nardiello
}
\note{
The package is still an MVP and very lightweight. Developers are welcome to participate in the project and contribute via GitHub.
}

\seealso{
\code{\link{abc_analysis}}, \code{\link{eoq}}, \code{\link{safety_stock}}, \code{\link{transport_optimization}}
}